requirements.txt
fastapi
uvicorn
python-multipart
baidu-aip
requests
beautifulsoup4